pyproject.toml
[project]
name = "cortex-gateway"
version = "0.1.0"
description = "Unified REST API gateway for Cortex email automation services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "cortex-utils @ git+https://github.com/devonjones/cortex-utils.git@main",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "isort>=5.13.0",
    "types-psycopg2>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gateway"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "gateway.blueprints.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
